requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
python-telegram-bot>=20.0
yookassa>=2.3.0
django-crispy-forms>=2.0
django-allauth>=0.54.0
django-ckeditor>=6.5.1
django-debug-toolbar>=4.0.0
django-environ>=0.10.0
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
whitenoise>=6.5.0